+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/button/ButtonPanelBuilder.java
@@ -19,13 +19,9 @@
 package is.codion.swing.common.ui.component.button;
 
 import is.codion.swing.common.ui.control.Controls;
-import is.codion.swing.common.ui.control.Controls.ControlsBuilder;
 
-import javax.swing.Action;
 import javax.swing.ButtonGroup;
 import javax.swing.JPanel;
-
-import static java.util.Objects.requireNonNull;
 
 /**
  * Builds a JPanel with buttons.
@@ -58,29 +54,5 @@ public interface ButtonPanelBuilder extends ControlPanelBuilder<JPanel, ButtonPa
 	 */
 	static ButtonPanelBuilder builder() {
 		return new DefaultButtonPanelBuilder((Controls) null);
-	}
-
-	/**
-	 * @param actions the actions
-	 * @return a new button panel builder
-	 */
-	static ButtonPanelBuilder builder(Action... actions) {
-		return new DefaultButtonPanelBuilder(requireNonNull(actions));
-	}
-
-	/**
-	 * @param controls the controls
-	 * @return a new button panel builder
-	 */
-	static ButtonPanelBuilder builder(Controls controls) {
-		return new DefaultButtonPanelBuilder(requireNonNull(controls));
-	}
-
-	/**
-	 * @param controlsBuilder the controls builder
-	 * @return a new button panel builder
-	 */
-	static ButtonPanelBuilder builder(ControlsBuilder controlsBuilder) {
-		return builder(requireNonNull(controlsBuilder).build());
 	}
 }